+++ src/proj6BittingCerratoCohenEllmer/view/Main.fxml
@@ -13,14 +13,13 @@
 <?import javafx.scene.control.MenuBar?>
 <?import javafx.scene.control.MenuItem?>
 <?import javafx.scene.control.SeparatorMenuItem?>
-<?import javafx.scene.control.TabPane?>
 <?import javafx.scene.control.ToolBar?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import org.fxmisc.richtext.StyleClassedTextArea?>
 
 <AnchorPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/10.0.2"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="proj6BittingCerratoCohenEllmer.Controller">
+            fx:controller="proj6BittingCerratoCohenEllmer.controllers.MasterController">
     <!-- make menu bar expand to fill the window if user resizes window  -->
     <MenuBar AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
              AnchorPane.topAnchor="0.0">
@@ -61,12 +60,11 @@
                       mnemonicParsing="false" onAction="#handleSelectAll"/>
         </Menu>
     </MenuBar>
-    <!-- use a TabPane to hold the text area  -->
-    <TabPane fx:id="tabPane" layoutX="14.0" layoutY="112.0"
-             prefHeight="396.0" prefWidth="800.0" tabClosingPolicy="ALL_TABS"
-             AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
-             AnchorPane.topAnchor="74.0">
-    </TabPane>
+    <!-- use a separate FXML file to enable nested controllers  -->
+    <!-- it is necessary to have an fx:id equal to the field in the main controller minus "MasterController"  -->
+    <!-- ex. in MasterController.java, the controller for the element below is named tabController  -->
+    <!--     here the fx:id should be "tab" -->
+    <fx:include fx:id="tab" source="CodeTabs.fxml"/>
     <!-- use a ToolBar to hold the two buttons -->
     <ToolBar layoutY="27.0" prefHeight="49.0" prefWidth="400.0"
              AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
